+++ CHANGELOG.md
@@ -31,6 +31,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   with Manim's.
   [#499](https://github.com/jeertmans/manim-slides/pull/499)
 
+- Bump ManimGL to `>=1.7.2`, to remove `pyrr` from dependencies,
+  and to avoid complex code for supporting both `1.7.1` and `>=1.7.2`,
+  as the latter includes many breaking changes.
+  [#499](https://github.com/jeertmans/manim-slides/pull/506)
+
 (v5.1.10)=
 ## [v5.1.10](https://github.com/jeertmans/manim-slides/compare/v5.1.9...v5.1.10)